\name{checkForIntercept.default}
\alias{checkForIntercept.default}
\title{checkForIntercept.default}
\usage{
  \method{checkForIntercept}{default} (model, ...)
}
\arguments{
  \item{model}{Model to be checked for an intercept}
}
\value{
  TRUE or FALSE (actually 0 or 1)
}
\description{
  Default method
}
\details{
  Check for an intercept in most models
}
\author{
  Jared P. Lander
}
\seealso{
  checkForIntercept
}

