% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot-package.r
\docType{package}
\name{coefplot-package}
\alias{coefplot-package}
\title{coefplot: Plots Coefficients from Fitted Models}
\description{
Plots the coefficients from model objects. This very quickly shows the user the point estimates and confidence intervals for fitted models.
}
\author{
\strong{Maintainer}: Jared Lander \email{packages@jaredlander.com} (\href{https://orcid.org/0000-0002-7291-726X}{ORCID})

}
\keyword{internal}
