% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.r
\name{coefplot.logitmfx}
\alias{coefplot.logitmfx}
\title{coefplot.logitmfx}
\usage{
\method{coefplot}{logitmfx}(...)
}
\arguments{
\item{\dots}{All arguments are passed on to \code{\link{coefplot.default}}.  Please see that function for argument information.}
}
\value{
A ggplot object.  See \code{\link{coefplot.lm}} for more information.
}
\description{
Dotplot for logitmfx coefficients
}
\details{
A graphical display of the coefficients and standard errors from a fitted logitmfx model

\code{\link{coefplot}} is the S3 generic method for plotting the coefficients from a fitted model.

For more information on this function and it's arguments see \code{\link{coefplot.default}}
}
\examples{

data(diamonds, package='ggplot2')
library(coefplot)
library(mfx)
margin1 <- logitmfx(price > 18000 ~ cut + carat, data=diamonds, 
atmean=FALSE, robust=TRUE)
coefplot(margin1)

}
\author{
Jared P. Lander
}

