\name{coefplot.mer}
\alias{coefplot.mer}
\title{coefplot.mer}
\usage{
  \method{coefplot}{mer} (model,
    effects = c("ranef", "fixef"), grouping = 1, ...)
}
\arguments{
  \item{model}{Model to be plotted}

  \item{effects}{Which type of effects to plot, only ranef
  implemented now}

  \item{grouping}{Which grouping factor t0 plot}

  \item{\dots}{Further arguments}
}
\value{
  A coefficient plot
}
\description{
  Plot Coefficients from lmer
}
\details{
  Details to follow
}
\author{
  Jared P. Lander
}

