% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.r
\name{coefplot.workflow}
\alias{coefplot.workflow}
\title{coefplot.workflow}
\usage{
\method{coefplot}{workflow}(model, ...)
}
\arguments{
\item{model}{A workflow object}

\item{\dots}{All arguments are passed on to \code{\link{coefplot.default}}.  Please see that function for argument information.}
}
\description{
Coefplot method for workflow objects
}
\details{
Pulls model element out of workflow object then calls \code{coefplot}.
}
\author{
Jared P. Lander
}
