% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCoef.r
\name{extract.coef.logitmfx}
\alias{extract.coef.logitmfx}
\title{extract.coef.logitmfx}
\usage{
\method{extract.coef}{logitmfx}(model, ...)
}
\arguments{
\item{model}{Model object to extract information from.}

\item{...}{Further arguments}
}
\value{
A \code{\link{data.frame}} containing the coefficient, the standard error and the variable name.
}
\description{
Extract Coefficient Information from logitmfx Models
}
\details{
Gets the coefficient values and standard errors, and variable names from an logitmfx model.
}
\examples{
\dontrun{
data(diamonds, package='ggplot2')
library(coefplot)
library(mfx)
margin1 <- logitmfx(price > 18000 ~ cut + carat, data=diamonds, 
atmean=FALSE, robust=TRUE)
extract.coef(margin1)
}

}
\author{
Jared P. Lander
}

