\name{get.assign.coxph}
\alias{get.assign.coxph}
\title{get.assign.coxph}
\usage{
  get.assign.coxph(model, ...)
}
\arguments{
  \item{model}{Fitted model}

  \item{\dots}{Further arguments}
}
\value{
  The assignment vector
}
\description{
  The assignment vector for an coxph model
}
\details{
  Gets relative positions of predictors
}
\author{
  Jared P. Lander
}

