\name{get.assign.glm}
\alias{get.assign.glm}
\title{get.assign.glm}
\usage{
  get.assign.glm(model, ...)
}
\arguments{
  \item{model}{Fitted model}

  \item{\dots}{Further arguments}
}
\value{
  The assignment vector
}
\description{
  The assignment vector for a glm model
}
\details{
  Gets relative positions of predictors
}
\author{
  Jared P. Lander
}

