% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.assign.lm}
\alias{get.assign.lm}
\title{get.assign.lm}
\usage{
\method{get.assign}{lm}(model, ...)
}
\arguments{
\item{model}{Fitted model}

\item{\dots}{Further arguments}
}
\value{
The assignment vector
}
\description{
The assignment vector for an lm model
}
\details{
Gets relative positions of predictors
}
\author{
Jared P. Lander
}

