% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefPredMatching.r
\name{getCoefsFromPredictors.default}
\alias{getCoefsFromPredictors.default}
\title{getCoefsFromPredictors.default}
\usage{
\method{getCoefsFromPredictors}{default}(model, predictors = NULL, strict = FALSE, ...)
}
\arguments{
\item{model}{A fitted model}

\item{predictors}{A character vector of predictors to match against.  Interactions can be explicitly specified by VariableA:VariableB.}

\item{strict}{Logical specifying if interactions terms should be included (\code{FALSE}) or just the main terms (\code{TRUE}).}

\item{\dots}{further arguments}
}
\value{
A character vector of coefficients listing the coefficients that match the predictor
}
\description{
Default function (lm, glm) for matching coefficients with predictors
}
\details{
The user specifies predictors whose coefficients should be included in the coefplot.
}
\author{
Jared P. Lander
}
