\name{getModelInfo.mer.ranef}
\alias{getModelInfo.mer.ranef}
\alias{getModelInfo.met}
\title{getModelInfo.mer}
\usage{
  getModelInfo.mer.ranef(model, grouping = 1, ...)
}
\arguments{
  \item{effects}{Indicates if we are plotting fixed
  (individual level) or random (group) effects}

  \item{model}{The fitted model with coefficients to be
  plotted}
}
\description{
  Title
}
\details{
  Details
}

