% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.r
\name{invlogit}
\alias{invlogit}
\title{invlogit}
\usage{
invlogit(x)
}
\arguments{
\item{x}{Vector of numbers}
}
\value{
\code{x} mapped to [0, 1]
}
\description{
Calculates the inverse logit
}
\details{
Maps the real line to [0, 1]
}
\examples{
invlogit(3)
invlogit(-6:6)
invlogit(c(-1, 1, 2))

}
\author{
Jared P. Lander
}
