% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefPredMatching.r
\name{matchCoefs.default}
\alias{matchCoefs.default}
\title{matchCoefs.default}
\usage{
\method{matchCoefs}{default}(model, ...)
}
\arguments{
\item{model}{Fitted model}

\item{\dots}{Further arguments}
}
\value{
a data.frame matching predictors to coefficients
}
\description{
Match coefficients to predictors
}
\details{
Matches coefficients to predictors using information from model matrices
}
\author{
Jared P. Lander
}
