% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPlottingFrame.r
\name{buildModelCI}
\alias{buildModelCI}
\title{buildModelCI}
\usage{
buildModelCI(model, ...)
}
\arguments{
\item{model}{A Fitted model such as from lm, glm}

\item{\dots}{Arguments passed on onto other methods}
}
\value{
A \code{\link{data.frame}} listing coefficients and confidence bands.
}
\description{
Construct Confidence Interval Values
}
\details{
Takes a model and builds a data.frame holding the coefficient value and the confidence interval values.
}
\examples{

data(diamonds)
model1 <- lm(price ~ carat + cut, data=diamonds)
coefplot:::buildModelCI(model1)
coefplot(model1)

}
\seealso{
\code{\link{coefplot}} \code{\link{multiplot}}
}
\author{
Jared P. Lander
}
