% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.r
\name{coefplot.rxLogit}
\alias{coefplot.rxLogit}
\title{coefplot.rxLogit}
\usage{
\method{coefplot}{rxLogit}(...)
}
\arguments{
\item{\dots}{All arguments are passed on to \code{\link{coefplot.lm}}.  Please see that function for argument information.}
}
\value{
A ggplot object.  See \code{\link{coefplot.lm}} for more information.
}
\description{
Dotplot for rxLogit coefficients
}
\details{
A graphical display of the coefficients and standard errors from a fitted rxLogit model

\code{\link{coefplot}} is the S3 generic method for plotting the coefficients from a fitted model.

For more information on this function and it's arguments see \code{\link{coefplot.lm}}
}
\examples{

\dontrun{
data(diamonds)
mod6 <- rxLogit(price > 10000 ~ carat + cut + x, data=diamonds)
coefplot(mod6)
}
}
\author{
Jared P. Lander www.jaredlander.com
}
