% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefplot.r
\name{coefplot.workflow}
\alias{coefplot.workflow}
\title{coefplot.workflow}
\usage{
\method{coefplot}{workflow}(model, ...)
}
\arguments{
\item{model}{A workflow object}

\item{\dots}{All arguments are passed on to \code{\link{coefplot.default}}.  Please see that function for argument information.}
}
\description{
Coefplot method for workflow objects
}
\details{
Pulls model element out of workflow object then calls \code{coefplot}.
}
\examples{

library(workflows)
lm_spec <- parsnip::linear_reg() \%>\% parsnip::set_engine('lm')
flow <- workflow() \%>\% 
    add_model(lm_spec) \%>\% 
    add_formula(mpg ~ cyl + qsec + disp)
mod5 <- parsnip::fit(flow, data=mtcars)# mod5 <- parsnip::fit(flow, data=mtcars)
coefplot(mod5)
coefplot(mod5, interactive=TRUE)
}
\author{
Jared P. Lander
}
