% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCoef.r
\name{extract.coef.maxLik}
\alias{extract.coef.maxLik}
\title{extract.coef.maxLik}
\usage{
\method{extract.coef}{maxLik}(model, ...)
}
\arguments{
\item{model}{Model object from which to extract information.}

\item{\dots}{Further arguments}
}
\value{
A \code{\link{data.frame}} containing the coefficient, the standard 
error and the variable name.
}
\description{
Extract Coefficient Information from Models
}
\details{
Gets the coefficient values and variable names from a model.
}
\author{
Jared P. Lander
}
