% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefPredMatching.r
\name{get.assign.glm}
\alias{get.assign.glm}
\title{get.assign.glm}
\usage{
\method{get.assign}{glm}(model, ...)
}
\arguments{
\item{model}{Fitted model}

\item{\dots}{Further arguments}
}
\value{
The assignment vector
}
\description{
The assignment vector for a glm model
}
\details{
Gets relative positions of predictors
}
\author{
Jared P. Lander
}
