% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coefPredMatching.r
\name{get.assign}
\alias{get.assign}
\title{get.assign}
\usage{
get.assign(model, ...)
}
\arguments{
\item{model}{Fitted model}

\item{\dots}{Further arguments}
}
\value{
The assignment vector
}
\description{
The assignment vector for a model
}
\details{
Gets relative positions of predictors
}
\author{
Jared P. Lander
}

