% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefPredMatching.r
\name{getCoefsFromPredictors}
\alias{getCoefsFromPredictors}
\title{getCoefsFromPredictors}
\usage{
getCoefsFromPredictors(model, predictors, ...)
}
\arguments{
\item{model}{A fitted model}

\item{predictors}{A character vector of predictors to match against}

\item{\dots}{further arguments}
}
\value{
A character vector of coefficients listing the coefficients that match the predictor
}
\description{
Generic function for finding which coefficients go with which predictors
}
\details{
The user specifies predictors whose coefficients should be included in the coefplot.
}
\author{
Jared P. Lander
}
