\name{position_dodgev}
\alias{position_dodgev}
\title{position_dodgev}
\usage{
  position_dodgev(width = NULL, height = NULL)
}
\arguments{
  \item{width}{}

  \item{height}{}
}
\value{
  Not sure
}
\description{
  Vertical dodging internal
}
\details{
  A hacky adaptation of ggplot's position_dodge function to
  be used for vertical collisions.  No warranties on this
  working except that it looks good for now.

  No changes were necessary but this is a new geom with a v
  on the end
}
\examples{
# none here
}
\author{
  Jared P. Lander
}

